# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorCode": {
        "enum": [
          "SCENE_NOT_FOUND",
          "JOB_NOT_FOUND",
          "CANDIDATE_NOT_FOUND",
          "ARTIFACT_NOT_FOUND",
          "ARMATURE_NOT_FOUND",
          "BONE_NOT_FOUND",
          "INVALID_FRAME_RANGE",
          "UNSUPPORTED_REPAIR_TYPE",
          "UNSUPPORTED_RIG",
          "INVALID_STATE",
          "NOT_READY",
          "IDEMPOTENCY_CONFLICT",
          "FILE_INVALID",
          "FILE_TOO_LARGE",
          "VALIDATION_ERROR",
          "NOT_IMPLEMENTED",
          "PLAN_INVALID",
          "PROVIDER_UNAVAILABLE",
          "PROVIDER_TIMEOUT",
          "PROVIDER_RATE_LIMITED",
          "WORKER_CRASHED",
          "WORKER_TIMEOUT",
          "WORKER_OUTPUT_INVALID",
          "RENDER_FAILED",
          "NO_VIABLE_CANDIDATE",
          "STORAGE_ERROR",
          "INTERNAL"
        ],
        "title": "ErrorCode",
        "type": "string"
      },
...
      "Problem": {
        "additionalProperties": false,
        "description": "An RFC 7807 problem detail, extended with a machine-readable ``code``.",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "errors": {
            "default": [],
            "items": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            },
            "title": "Errors",
            "type": "array"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instance"
          },
          "status": {
            "maximum": 599.0,
            "minimum": 400.0,
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "title",
          "status",
          "code"
        ],
        "title": "Problem",
        "type": "object"
      },
      "ProductStats": {
        "additionalProperties": false,
        "description": "Product metrics computed from the store (docs/METRICS.md).",
        "properties": {
          "candidate_acceptance_rate": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Candidate Acceptance Rate"
          },
          "cost_per_accepted_repair_usd": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost Per Accepted Repair Usd"
          },
          "jobs_decided": {
            "minimum": 0.0,
            "title": "Jobs Decided",
            "type": "integer"
          },
          "jobs_total": {
            "minimum": 0.0,
            "title": "Jobs Total",
            "type": "integer"
          },
          "median_slip_reduction_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Median Slip Reduction Pct"
          },
          "median_time_to_decision_s": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Median Time To Decision S"
          },
          "model_human_agreement_rate": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Human Agreement Rate"
          }
        },
        "required": [
          "jobs_total",
          "jobs_decided"
        ],
        "title": "ProductStats",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "AI plans and evaluates; deterministic code executes; the animator decides. Every error body is an RFC 7807 Problem carrying a machine-readable `code`.",
    "summary": "Localized animation repair: measure, propose, compare, apply.",
    "title": "Kinesis API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/stats/product": {
      "get": {
        "operationId": "product_stats_v1_stats_product_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductStats"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Problem"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Problem"
          }
        },
        "summary": "Product metrics (acceptance, agreement, slip reduction, cost)",
        "tags": [
          "ops"
        ]
      }
    }
  }
}